Add support for the TTN packet fowarder
Can no be used with Kersing's poly packet forwarder, mp packet
forwarder, or the TTN's packet forwarder.

Clean up the configuration and files by versions of the packet
forwarder not installed.

If specified in the config file, the TTN router to be used is
configured on the TTN.  It is necessary to configure this for the TTN
packet forwarder so it's better to configure it here.

The 'register' tag has been merged into the 'loraconfig' tag.