More improvements for interaction with TTN
The register command now handles all updates to TTN.

This avoids errors observed when trying to run multiple copies of
ttnctl at once.  Register uses locking to serialize requests.

Register will also spit out the gateway info as json for use in
templates.

It would be nice to update register to use the REST API directly as
this would be more efficent and allow updating and returning of more
info.