Package cleanup
Nodejs is only required when not using mp_packet_forwarder 3.0.25 or
later.  Don't try to install if we do not need it (it's not available
on 5.3.0 or later).  Remove it if we don't need it.

In 5.3.0 and later some python packages are not available, ignore
errors trying to install them.

Run an `opkg clean` after changing packages.  Ignore errors because
that is a new feature in mLinux 5.