Fix mlinux-firmware-upgrade, do not copy to target
The mlinux-firmware-upgrade would leave it's temporary directory on
the Ansible controller because TMPDIR is relative and it could be in
TMPDIR when trying to remove it.

Our version of mlinux-firmware-upgrade comes from an older version of
mLinux.  We keep it around so we can run it on the Ansible controller
to validate a firmware upgrade file.  But we only want to copy it to
the target if there is no version already on the target.