Add a reset in the basic station startup script
Apparently the concentrator can get hung up such that even a power
cycle will not fix it.  Only these reset commands.  We had them with
the mp and poly pkt forwarders and switching back to that from basic
station fixed the problem.  So let's put them in the basic-station
script too.